requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.35.0
azure-storage-blob==12.19.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
marshmallow==3.21.2
openai==1.96.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3